backend/pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Andrej Chikilev <andrej.chikilev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pyyaml = "^6.0.2"
uritemplate = "^4.1.1"
inflection = "^0.5.1"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
